refactor: move passport strategies into src folder #591
